refactor vimstate, recordedstate to own files
