+++ src/actions/commands/insert.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
-import { RecordedState, VimState } from './../../mode/modeHandler';
+import { RecordedState } from '../../state/recordedState';
+import { VimState } from '../../state/vimState';
 import { Register, RegisterMode } from './../../register/register';
 import { Position, PositionDiff } from './../../common/motion/position';
 import { Range } from './../../common/motion/range';
