+++ src/actions/plugins/easymotion/easymotion.cmd.ts
@@ -4,7 +4,7 @@ import { ModeName } from './../../../mode/mode';
 import { Configuration } from './../../../configuration/configuration';
 import { BaseCommand } from './../../commands/actions';
 import { RegisterAction } from './../../base';
-import { VimState } from './../../../mode/modeHandler';
+import { VimState } from '../../../state/vimState';
 import {
   EasyMotionMoveOptionsBase,
   EasyMotionWordMoveOpions,
