+++ src/mode/modeHandler.ts
@@ -3,26 +3,24 @@ import { SurroundInputMode } from './surroundInputMode';
 import * as vscode from 'vscode';
 import * as _ from 'lodash';
 import Globals from '../globals';
-import EditorIdentity from './../editorIdentity';
 
 import {
   isTextTransformation,
   TextTransformations,
   areAnyTransformationsOverlapping,
-  Transformation,
 } from './../transformations/transformations';
 import { Mode, ModeName, VSCodeVimCursorType } from './mode';
 import { InsertModeRemapper, OtherModesRemapper } from './remapper';
 import { NormalMode } from './modeNormal';
 import { InsertMode } from './modeInsert';
+import { ReplaceMode } from './modeReplace';
 import { VisualBlockMode } from './modeVisualBlock';
 import { VisualMode } from './modeVisual';
-import { taskQueue } from './../taskQueue';
-import { ReplaceMode } from './modeReplace';
+import { VisualLineMode } from './modeVisualLine';
 import { EasyMotionMode, EasyMotionInputMode } from './modeEasyMotion';
 import { SearchInProgressMode } from './modeSearchInProgress';
+import { taskQueue } from './../taskQueue';
 import { TextEditor } from './../textEditor';
-import { VisualLineMode } from './modeVisualLine';
 import { HistoryTracker } from './../history/historyTracker';
 import { EasyMotion } from './../actions/plugins/easymotion/easymotion';
 import { Actions, KeypressState, BaseAction } from './../actions/base';
@@ -38,460 +36,13 @@ import { Position, PositionDiff } from './../common/motion/position';
 import { Range } from './../common/motion/range';
 import { RegisterMode, Register } from './../register/register';
 import { showCmdLine } from '../../src/cmd_line/main';
-import { Configuration } from '../../src/configuration/configuration';
+import { Configuration } from '../configuration/configuration';
 import { PairMatcher } from './../common/matching/matcher';
 import { ReplaceState } from './../state/replaceState';
 import { GlobalState } from './../state/globalState';
-import { Nvim } from 'promised-neovim-client';
+import { RecordedState } from './../state/recordedState';
+import { VimState } from './../state/vimState';
 import { allowVSCodeToPropagateCursorUpdatesAndReturnThem } from '../util';
-
-export class ViewChange {
-  public command: string;
-  public args: any;
-}
-
-/**
- * The VimState class holds permanent state that carries over from action
- * to action.
- *
- * Actions defined in actions.ts are only allowed to mutate a VimState in order to
- * indicate what they want to do.
- */
-export class VimState {
-  /**
-   * The column the cursor wants to be at, or Number.POSITIVE_INFINITY if it should always
-   * be the rightmost column.
-   *
-   * Example: If you go to the end of a 20 character column, this value
-   * will be 20, even if you press j and the next column is only 5 characters.
-   * This is because if the third column is 25 characters, the cursor will go
-   * back to the 20th column.
-   */
-  public desiredColumn = 0;
-
-  public historyTracker: HistoryTracker;
-
-  public easyMotion: EasyMotion;
-
-  /**
-   * Just for debugging!!
-   */
-  public identity: EditorIdentity;
-
-  public editor: vscode.TextEditor;
-
-  /**
-   * For timing out remapped keys like jj to esc.
-   */
-  public lastKeyPressedTimestamp = 0;
-
-  /**
-   * Are multiple cursors currently present?
-   */
-  public isMultiCursor = false;
-
-  // Is the multicursor something like visual block "multicursor", where
-  // natively in vim there would only be one cursor whose changes were applied
-  // to all lines after edit.
-  public isFakeMultiCursor = false;
-
-  /**
-   * Tracks movements that can be repeated with ; and , (namely t, T, f, and F).
-   */
-  public static lastRepeatableMovement: BaseMovement | undefined = undefined;
-
-  public lastMovementFailed: boolean = false;
-
-  public alteredHistory = false;
-
-  public isRunningDotCommand = false;
-
-  public focusChanged = false;
-
-  public surround:
-    | undefined
-    | {
-        active: boolean;
-        operator: 'change' | 'delete' | 'yank';
-        target: string | undefined;
-        replacement: string | undefined;
-        range: Range | undefined;
-        isVisualLine: boolean;
-      } = undefined;
-
-  /**
-   * Used for command like <C-o> which allows you to return to insert after a command
-   */
-  public returnToInsertAfterCommand = false;
-  public actionCount = 0;
-
-  /**
-   * Every time we invoke a VS Code command which might trigger Code's view update,
-   * we should postpone its view updating phase to avoid conflicting with our internal view updating mechanism.
-   * This array is used to cache every VS Code view updating event and they will be triggered once we run the inhouse `viewUpdate`.
-   */
-  public postponedCodeViewChanges: ViewChange[] = [];
-
-  /**
-   * Used to prevent non-recursive remappings from looping.
-   */
-  public isCurrentlyPerformingRemapping = false;
-
-  /**
-   * All the keys we've pressed so far.
-   */
-  public keyHistory: string[] = [];
-
-  public globalState: GlobalState = new GlobalState();
-
-  /**
-   * The position the cursor will be when this action finishes.
-   */
-  public get cursorPosition(): Position {
-    return this.allCursors[0].stop;
-  }
-  public set cursorPosition(value: Position) {
-    this.allCursors[0] = this.allCursors[0].withNewStop(value);
-  }
-
-  /**
-   * The effective starting position of the movement, used along with cursorPosition to determine
-   * the range over which to run an Operator. May rarely be different than where the cursor
-   * actually starts e.g. if you use the "aw" text motion in the middle of a word.
-   */
-  public get cursorStartPosition(): Position {
-    return this.allCursors[0].start;
-  }
-  public set cursorStartPosition(value: Position) {
-    this.allCursors[0] = this.allCursors[0].withNewStart(value);
-  }
-
-  /**
-   * In Multi Cursor Mode, the position of every cursor.
-   */
-  private _allCursors: Range[] = [new Range(new Position(0, 0), new Position(0, 0))];
-
-  public get allCursors(): Range[] {
-    return this._allCursors;
-  }
-
-  public set allCursors(value: Range[]) {
-    for (const cursor of value) {
-      if (!cursor.start.isValid() || !cursor.stop.isValid()) {
-        console.log('invalid value for set cursor position. This is probably bad?');
-      }
-    }
-
-    this._allCursors = value;
-
-    this.isMultiCursor = this._allCursors.length > 1;
-  }
-
-  public cursorPositionJustBeforeAnythingHappened = [new Position(0, 0)];
-
-  public isRecordingMacro: boolean = false;
-  public isReplayingMacro: boolean = false;
-
-  public replaceState: ReplaceState | undefined = undefined;
-
-  /**
-   * Stores last visual mode for gv
-   */
-  public lastVisualMode: ModeName;
-
-  /**
-   * Last selection that was active
-   */
-  public lastVisualSelectionStart: Position;
-  public lastVisualSelectionEnd: Position;
-
-  /**
-   * Was the previous mouse click past EOL
-   */
-  public lastClickWasPastEol: boolean = false;
-
-  /**
-   * The mode Vim will be in once this action finishes.
-   */
-  private _currentMode: ModeName;
-
-  public get currentMode(): number {
-    return this._currentMode;
-  }
-
-  public set currentMode(value: number) {
-    this._currentMode = value;
-
-    vscode.commands.executeCommand('setContext', 'vim.mode', ModeName[value]);
-  }
-
-  public currentModeName(): string {
-    return ModeName[this._currentMode];
-  }
-
-  public getModeObject(modeHandler: ModeHandler): Mode {
-    return modeHandler.modeList.find(mode => mode.isActive)!;
-  }
-
-  public currentRegisterMode = RegisterMode.FigureItOutFromCurrentMode;
-
-  public effectiveRegisterMode(): RegisterMode {
-    if (this.currentRegisterMode === RegisterMode.FigureItOutFromCurrentMode) {
-      if (this.currentMode === ModeName.VisualLine) {
-        return RegisterMode.LineWise;
-      } else if (this.currentMode === ModeName.VisualBlock) {
-        return RegisterMode.BlockWise;
-      } else {
-        return RegisterMode.CharacterWise;
-      }
-    } else {
-      return this.currentRegisterMode;
-    }
-  }
-
-  /**
-   * The top left of a selected block of text. Useful for Visual Block mode.
-   */
-  public get topLeft(): Position {
-    return VisualBlockMode.getTopLeftPosition(this.cursorStartPosition, this.cursorPosition);
-  }
-
-  /**
-   * The bottom right of a selected block of text. Useful for Visual Block mode.
-   */
-  public get bottomRight(): Position {
-    return VisualBlockMode.getBottomRightPosition(this.cursorStartPosition, this.cursorPosition);
-  }
-
-  public registerName = '"';
-
-  public commandInitialText = '';
-
-  public recordedState = new RecordedState();
-
-  public recordedMacro = new RecordedState();
-
-  /**
-   * Programmatically triggering an edit will unfortunately ALSO trigger our mouse update
-   * function. We use this variable to determine if the update function was triggered
-   * by us or by a mouse action.
-   */
-  public whatILastSetTheSelectionTo: vscode.Selection;
-
-  public nvim: Nvim;
-
-  public constructor(editor: vscode.TextEditor, startInInsertMode: boolean) {
-    this.editor = editor;
-    this.identity = new EditorIdentity(editor);
-    this.historyTracker = new HistoryTracker(this);
-    this.easyMotion = new EasyMotion();
-
-    this.currentMode = startInInsertMode ? ModeName.Insert : ModeName.Normal;
-  }
-}
-
-/**
- * The RecordedState class holds the current action that the user is
- * doing. Example: Imagine that the user types:
- *
- * 5"qdw
- *
- * Then the relevent state would be
- *   * count of 5
- *   * copy into q register
- *   * delete operator
- *   * word movement
- *
- *
- * Or imagine the user types:
- *
- * vw$}}d
- *
- * Then the state would be
- *   * Visual mode action
- *   * (a list of all the motions you ran)
- *   * delete operator
- */
-export class RecordedState {
-  constructor() {
-    this.registerName = Configuration.useSystemClipboard ? '*' : '"';
-  }
-
-  /**
-   * The keys the user has pressed that have not caused an action to be
-   * executed yet. Used for showcmd and command remapping.
-   */
-  public commandList: string[] = [];
-
-  /**
-   * The number of keys the user has pressed that have been remapped.
-   */
-  public numberOfRemappedKeys: number = 0;
-
-  /**
-   * String representation of the exact keys that the user entered. Used for
-   * showcmd.
-   */
-  public get commandString(): string {
-    let result = '';
-
-    for (const key of this.commandList) {
-      if (key === Configuration.leader) {
-        result += '<leader>';
-      } else {
-        result += key;
-      }
-    }
-
-    return result;
-  }
-  /**
-   * get the current command without the prefixed count.
-   * For instance: if the current commandList is ['2', 'h'], returns only ['h'].
-   */
-  public getCurrentCommandWithoutCountPrefix(): string[] {
-    const commandList = this.commandList;
-    const result = [];
-    let previousWasCount = true;
-
-    for (const commandKey of commandList) {
-      if (previousWasCount && commandKey.match(/[0-9]/)) {
-        continue;
-      } else {
-        previousWasCount = false;
-        result.push(commandKey);
-      }
-    }
-
-    return result;
-  }
-
-  /**
-   * lenth of the current command with remappings and the prefixed count excluded.
-   */
-  public get numberOfKeysInCommandWithoutCountPrefix() {
-    return this.getCurrentCommandWithoutCountPrefix().length - this.numberOfRemappedKeys;
-  }
-
-  /**
-   * Reset the command list.
-   */
-  public resetCommandList() {
-    this.commandList = [];
-    this.numberOfRemappedKeys = 0;
-  }
-
-  /**
-   * Keeps track of keys pressed for the next action. Comes in handy when parsing
-   * multiple length movements, e.g. gg.
-   */
-  public actionKeys: string[] = [];
-
-  /**
-   * Every action that has been run.
-   */
-  public actionsRun: BaseAction[] = [];
-
-  public hasRunOperator = false;
-
-  public hasRunSurround = false;
-  public surroundKeys: string[] = [];
-  public surroundKeyIndexStart = 0;
-
-  /**
-   * This is kind of a hack and should be associated with something like this:
-   *
-   * https://github.com/VSCodeVim/Vim/issues/805
-   */
-  public operatorPositionDiff: PositionDiff | undefined;
-
-  public isInsertion = false;
-
-  /**
-   * The text transformations that we want to run. They will all be run after the action has been processed.
-   *
-   * Running an individual action will generally queue up to one of these, but if you're in
-   * multi-cursor mode, you'll queue one per cursor, or more.
-   *
-   * Note that the text transformations are run in parallel. This is useful in most cases,
-   * but will get you in trouble in others.
-   */
-  public transformations: Transformation[] = [];
-
-  /**
-   * The operator (e.g. d, y, >) the user wants to run, if there is one.
-   */
-  public get operator(): BaseOperator {
-    let list = _.filter(this.actionsRun, a => a instanceof BaseOperator).reverse();
-    return list[0] as any;
-  }
-
-  public get operators(): BaseOperator[] {
-    return _.filter(this.actionsRun, a => a instanceof BaseOperator).reverse() as any;
-  }
-
-  /**
-   * The command (e.g. i, ., R, /) the user wants to run, if there is one.
-   */
-  public get command(): BaseCommand {
-    const list = _.filter(this.actionsRun, a => a instanceof BaseCommand);
-
-    // TODO - disregard <Esc>, then assert this is of length 1.
-
-    return list[0] as any;
-  }
-
-  public get hasRunAMovement(): boolean {
-    return _.filter(this.actionsRun, a => a.isMotion).length > 0;
-  }
-
-  /**
-   * The number of times the user wants to repeat this action.
-   */
-  public count: number = 0;
-
-  /**
-   * The register name for this action.
-   */
-  public registerName: string;
-
-  public clone(): RecordedState {
-    const res = new RecordedState();
-
-    // TODO: Actual clone.
-
-    res.actionKeys = this.actionKeys.slice(0);
-    res.actionsRun = this.actionsRun.slice(0);
-    res.hasRunOperator = this.hasRunOperator;
-    res.hasRunSurround = this.hasRunSurround;
-    res.surroundKeys = this.surroundKeys;
-
-    return res;
-  }
-
-  public operatorReadyToExecute(mode: ModeName): boolean {
-    // Visual modes do not require a motion -- they ARE the motion.
-    return (
-      this.operator &&
-      !this.hasRunOperator &&
-      mode !== ModeName.SearchInProgressMode &&
-      (this.hasRunAMovement ||
-        (mode === ModeName.Visual || mode === ModeName.VisualLine) ||
-        (this.operators.length > 1 &&
-          this.operators.reverse()[0].constructor === this.operators.reverse()[1].constructor))
-    );
-  }
-
-  public get isInInitialState(): boolean {
-    return this.operator === undefined && this.actionsRun.length === 0 && this.count === 1;
-  }
-}
-
-interface IViewState {
-  selectionStart: Position;
-  selectionStop: Position;
-  currentMode: ModeName;
-}
 
 export class ModeHandler implements vscode.Disposable {
   private _disposables: vscode.Disposable[] = [];
@@ -507,7 +58,7 @@ export class ModeHandler implements vscode.Disposable {
     return this._vimState;
   }
 
-  private _caretDecoration = vscode.window.createTextEditorDecorationType({
+  private readonly _caretDecoration = vscode.window.createTextEditorDecorationType({
     backgroundColor: new vscode.ThemeColor('editorCursor.foreground'),
     borderColor: new vscode.ThemeColor('editorCursor.foreground'),
     dark: {
