+++ src/register/register.ts
@@ -1,4 +1,5 @@
-import { VimState, RecordedState } from './../mode/modeHandler';
+import { RecordedState } from './../state/recordedState';
+import { VimState } from './../state/vimState';
 import { YankOperator, BaseOperator, DeleteOperator } from './../actions/operator';
 import { CommandYankFullLine, BaseCommand, CommandRegister } from './../actions/commands/actions';
 import * as util from './../util';
